+++ packages/client/src/entities/user/constants/constants.ts
@@ -1,5 +1,31 @@
-// TODO: Переместить в папку entities/user/constants
-const INPUTS = [
+export const options = {
+  errorsContainer: '#second-error-field',
+  errorLabelStyle: {
+    fontSize: '0.8rem',
+    lineHeight: '1rem',
+    color: 'rgb(184, 50, 50)',
+  },
+  errorLabelCssClass: 'auth-error-label',
+}
+
+export const INPUTS_SIGIN = [
+  {
+    id: 1,
+    selector: 'login',
+    name: 'login',
+    type: 'text',
+    label: 'Логин',
+  },
+  {
+    id: 2,
+    selector: 'password',
+    name: 'password',
+    type: 'password',
+    label: 'Пароль',
+  },
+]
+
+export const INPUTS_SIGUP = [
   {
     id: 1,
     name: 'first_name',
@@ -50,5 +76,3 @@ const INPUTS = [
     label: 'Подтвердите пароль',
   },
 ]
-
-export { INPUTS }
