+++ packages/client/src/entities/user/lib/index.ts
@@ -1,2 +1,2 @@
-// Сюда импортируем\экспортируем утилиты сущности
-// export * from './configure-online-state'
+export * from './validation'
+export * from './handle-request'
