+++ packages/client/src/pages/register/register.tsx
@@ -1,13 +1,13 @@
 import { useEffect } from 'react'
 import AuthForm from '../../shared/ui/auth-form/auth-form'
-import { formValidation } from './model/formValidation'
-import { INPUTS } from './model/constants'
 import { Container, CssBaseline, Box } from '@mui/material'
 import styles from './styles'
+import { INPUTS_SIGUP } from '../../entities/user/constants'
+import { formValidationSignup } from '../../entities/user/lib'
 
 const Register = () => {
   useEffect(() => {
-    formValidation()
+    formValidationSignup()
   }, [])
 
   const link = {
@@ -21,7 +21,7 @@ const Register = () => {
       <Container maxWidth="sm" sx={styles.container}>
         <AuthForm
           id="auth-form"
-          inputs={INPUTS}
+          inputs={INPUTS_SIGUP}
           pageName="Регистрация"
           buttonText="Зарегистрироваться"
           link={link}
